requirements.txt
chainlit~=0.6.0
langchain~=0.0.240
python-dotenv~=1.0.0
pinecone-client
sentence-transformers
faiss-cpu
pymupdf
tiktoken
pytest